requirements.txt
annotated-types      ==0.6.0
anyio                ==3.7.1
certifi              ==2023.7.22
click                ==8.1.7
colorama             ==0.4.6
dnspython            ==2.4.2
email-validator      ==2.1.0.post1
fastapi              ==0.104.0
gunicorn             ==21.2.0
h11                  ==0.14.0
httpcore             ==0.18.0
httptools            ==0.6.1
httpx                ==0.25.0
idna                 ==3.4
itsdangerous         ==2.1.2
Jinja2               ==3.1.2
MarkupSafe           ==2.1.3
orjson               ==3.9.9
packaging            ==23.2
pip                  ==22.3.1
pydantic             ==2.4.2
pydantic_core        ==2.10.1
pydantic-extra-types ==2.1.0
pydantic-settings    ==2.0.3
pymongo              ==4.5.0
python-dotenv        ==1.0.0
python-multipart     ==0.0.6
PyYAML               ==6.0.1
setuptools           ==65.5.0
sniffio              ==1.3.0
starlette            ==0.27.0
typing_extensions    ==4.8.0
ujson               ==5.8.0
uvicorn              ==0.23.2
watchfiles           ==0.21.0
websockets           ==12.0
